refactor: move image verification handler out of webhooks package
