+++ pkg/webhooks/resource/handlers.go
@@ -2,7 +2,6 @@ package resource
 
 import (
 	"fmt"
-	"reflect"
 	"time"
 
 	"github.com/go-logr/logr"
@@ -13,9 +12,7 @@ import (
 	"github.com/kyverno/kyverno/pkg/clients/dclient"
 	"github.com/kyverno/kyverno/pkg/common"
 	"github.com/kyverno/kyverno/pkg/config"
-	"github.com/kyverno/kyverno/pkg/engine"
 	enginectx "github.com/kyverno/kyverno/pkg/engine/context"
-	"github.com/kyverno/kyverno/pkg/engine/response"
 	engineutils2 "github.com/kyverno/kyverno/pkg/engine/utils"
 	"github.com/kyverno/kyverno/pkg/event"
 	"github.com/kyverno/kyverno/pkg/metrics"
@@ -27,14 +24,12 @@ import (
 	jsonutils "github.com/kyverno/kyverno/pkg/utils/json"
 	"github.com/kyverno/kyverno/pkg/webhooks"
 	"github.com/kyverno/kyverno/pkg/webhooks/resource/audit"
+	"github.com/kyverno/kyverno/pkg/webhooks/resource/imageverification"
 	"github.com/kyverno/kyverno/pkg/webhooks/resource/mutation"
 	"github.com/kyverno/kyverno/pkg/webhooks/resource/validation"
 	webhookgenerate "github.com/kyverno/kyverno/pkg/webhooks/updaterequest"
 	webhookutils "github.com/kyverno/kyverno/pkg/webhooks/utils"
-	"github.com/pkg/errors"
 	admissionv1 "k8s.io/api/admission/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	corev1listers "k8s.io/client-go/listers/core/v1"
 	rbacv1listers "k8s.io/client-go/listers/rbac/v1"
 )
@@ -197,7 +192,8 @@ func (h *handlers) Mutate(logger logr.Logger, request *admissionv1.AdmissionRequ
 		return admissionutils.ResponseFailure(err.Error())
 	}
 	newRequest := patchRequest(mutatePatches, request, logger)
-	imagePatches, imageVerifyWarnings, err := h.applyImageVerifyPolicies(logger, newRequest, policyContext, verifyImagesPolicies)
+	ivh := imageverification.NewImageVerificationHandler(logger)
+	imagePatches, imageVerifyWarnings, err := ivh.Handle(h.metricsConfig, newRequest, verifyImagesPolicies, policyContext)
 	if err != nil {
 		logger.Error(err, "image verification failed")
 		return admissionutils.ResponseFailure(err.Error())
@@ -211,73 +207,6 @@ func (h *handlers) Mutate(logger logr.Logger, request *admissionv1.AdmissionRequ
 	admissionResponse := admissionutils.ResponseSuccessWithPatch(patch)
 	logger.V(4).Info("completed mutating webhook", "response", admissionResponse)
 	return admissionResponse
-}
-
-func (h *handlers) applyImageVerifyPolicies(logger logr.Logger, request *admissionv1.AdmissionRequest, policyContext *engine.PolicyContext, policies []kyvernov1.PolicyInterface) ([]byte, []string, error) {
-	ok, message, imagePatches, warnings := h.handleVerifyImages(logger, request, policyContext, policies)
-	if !ok {
-		return nil, nil, errors.New(message)
-	}
-
-	logger.V(6).Info("images verified", "patches", string(imagePatches), "warnings", warnings)
-	return imagePatches, warnings, nil
-}
-
-func (h *handlers) handleVerifyImages(logger logr.Logger, request *admissionv1.AdmissionRequest, policyContext *engine.PolicyContext, policies []kyvernov1.PolicyInterface) (bool, string, []byte, []string) {
-	if len(policies) == 0 {
-		return true, "", nil, nil
-	}
-
-	var engineResponses []*response.EngineResponse
-	var patches [][]byte
-	verifiedImageData := &engine.ImageVerificationMetadata{}
-	for _, p := range policies {
-		policyContext.Policy = p
-		resp, ivm := engine.VerifyAndPatchImages(policyContext)
-
-		engineResponses = append(engineResponses, resp)
-		patches = append(patches, resp.GetPatches()...)
-		verifiedImageData.Merge(ivm)
-	}
-
-	failurePolicy := policyContext.Policy.GetSpec().GetFailurePolicy()
-	blocked := webhookutils.BlockRequest(engineResponses, failurePolicy, logger)
-	if !isResourceDeleted(policyContext) {
-		events := webhookutils.GenerateEvents(engineResponses, blocked)
-		h.eventGen.Add(events...)
-	}
-
-	if blocked {
-		logger.V(4).Info("admission request blocked")
-		return false, webhookutils.GetBlockedMessages(engineResponses), nil, nil
-	}
-
-	prInfos := policyreport.GeneratePRsFromEngineResponse(engineResponses, logger)
-	h.prGenerator.Add(prInfos...)
-
-	if !verifiedImageData.IsEmpty() {
-		hasAnnotations := hasAnnotations(policyContext)
-		annotationPatches, err := verifiedImageData.Patches(hasAnnotations, logger)
-		if err != nil {
-			logger.Error(err, "failed to create image verification annotation patches")
-		} else {
-			// add annotation patches first
-			patches = append(annotationPatches, patches...)
-		}
-	}
-
-	warnings := webhookutils.GetWarningMessages(engineResponses)
-	return true, "", jsonutils.JoinPatches(patches...), warnings
-}
-
-func isResourceDeleted(policyContext *engine.PolicyContext) bool {
-	var deletionTimeStamp *metav1.Time
-	if reflect.DeepEqual(policyContext.NewResource, unstructured.Unstructured{}) {
-		deletionTimeStamp = policyContext.NewResource.GetDeletionTimestamp()
-	} else {
-		deletionTimeStamp = policyContext.OldResource.GetDeletionTimestamp()
-	}
-	return deletionTimeStamp != nil
 }
 
 func (h *handlers) handleDelete(logger logr.Logger, request *admissionv1.AdmissionRequest) {
